+++ src/add-ons/kernel/file_systems/layers/attribute_overlay/attribute_overlay.cpp
@@ -1212,7 +1212,7 @@ overlay_get_file_map(fs_volume *volume, fs_vnode *vnode, off_t offset,
 
 
 static status_t
-overlay_ioctl(fs_volume *volume, fs_vnode *vnode, void *cookie, ulong op,
+overlay_ioctl(fs_volume *volume, fs_vnode *vnode, void *cookie, uint32 op,
 	void *buffer, size_t length)
 {
 	OVERLAY_CALL(ioctl, cookie, op, buffer, length)
